+++ src/manage_sql_nvts.h
@@ -95,6 +95,12 @@
    { NULL, NULL, KEYWORD_TYPE_UNKNOWN }                                     \
  }
 
+const char *
+get_osp_vt_update_socket ();
+
+void
+set_osp_vt_update_socket (const char *new_socket);
+
 void
 check_db_nvts ();